# apps/backend/src/modules/auth/dto/resend-verification.dto.ts
import { ApiProperty } from "@nestjs/swagger";
import { IsEmail } from "class-validator";

export class ResendVerificationDto {
	@ApiProperty({ example: "user@example.com" })
	@IsEmail()
	email!: string;
}