# docs/swagger.json
{
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/auth/sign-up": {
			"post": {
				"description": "Sends a verification code to email. Use verify-email to complete registration.",
				"operationId": "AuthController_signUp",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/SignUpDto" }
						}
					}
				},
				"responses": {
					"201": { "description": "User created; check email for code" },
					"409": {
						"description": "Conflict - Resource already exists or state conflict",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [],
				"summary": "Register a new user",
				"tags": ["Auth"]
			}
		},
		"/api/v1/auth/verify-email": {
			"post": {
				"operationId": "AuthController_verifyEmail",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/VerifyEmailDto" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "Verified; tokens issued",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/AuthResponseSchema" }
							}
						}
					},
					"401": {
						"description": "Invalid or expired code",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [],
				"summary": "Verify email with 6-digit code",
				"tags": ["Auth"]
			}
		},
		"/api/v1/auth/resend-verification": {
			"post": {
				"operationId": "AuthController_resendVerification",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/ResendVerificationDto" }
						}
					}
				},
				"responses": { "200": { "description": "Generic success message" } },
				"security": [],
				"summary": "Resend verification code",
				"tags": ["Auth"]
			}
		},
		"/api/v1/auth/sign-in": {
			"post": {
				"description": "Requires verified email. Refresh token is set as httpOnly cookie.",
				"operationId": "AuthController_signIn",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/SignInDto" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successfully authenticated",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/AuthResponseSchema" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - Missing or invalid authentication",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"403": {
						"description": "Email not verified",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [],
				"summary": "Authenticate user",
				"tags": ["Auth"]
			}
		},
		"/api/v1/auth/forgot-password": {
			"post": {
				"operationId": "AuthController_forgotPassword",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/ForgotPasswordDto" }
						}
					}
				},
				"responses": {
					"200": { "description": "Generic message (no email enumeration)" }
				},
				"security": [],
				"summary": "Request password reset code",
				"tags": ["Auth"]
			}
		},
		"/api/v1/auth/reset-password": {
			"post": {
				"operationId": "AuthController_resetPassword",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/ResetPasswordDto" }
						}
					}
				},
				"responses": {
					"200": { "description": "Password reset" },
					"401": {
						"description": "Invalid or expired code",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [],
				"summary": "Reset password with code",
				"tags": ["Auth"]
			}
		},
		"/api/v1/auth/refresh": {
			"post": {
				"description": "Uses httpOnly refresh cookie; rotates refresh token.",
				"operationId": "AuthController_refresh",
				"parameters": [],
				"responses": {
					"200": {
						"description": "New access token",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/AuthResponseSchema" }
							}
						}
					},
					"401": {
						"description": "Invalid session",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [],
				"summary": "Refresh access token",
				"tags": ["Auth"]
			}
		},
...
		"/api/v1/auth/me": {
			"get": {
				"operationId": "AuthController_me",
				"parameters": [],
				"responses": {
					"200": { "description": "Current user profile" },
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [],
				"summary": "Current user",
				"tags": ["Auth"]
			}
		},
...
		"/api/v1/analytics/{endpointId}/summary": {
			"get": {
				"description": "Returns aggregated analytics for an endpoint (status counts, latency, etc.).",
				"operationId": "AnalyticsController_getSummary",
				"parameters": [
					{
						"name": "endpointId",
						"required": true,
						"in": "path",
						"description": "Endpoint UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "Analytics summary" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Get analytics summary",
				"tags": ["Analytics"]
			}
		},
		"/api/v1/analytics/{endpointId}/timeseries": {
			"get": {
				"description": "Returns time-bucketed analytics for charts.",
				"operationId": "AnalyticsController_getTimeseries",
				"parameters": [
					{
						"name": "endpointId",
						"required": true,
						"in": "path",
						"description": "Endpoint UUID",
						"schema": { "format": "uuid", "type": "string" }
					},
					{
						"name": "bucket",
						"required": false,
						"in": "query",
						"schema": { "enum": ["hour", "day"], "type": "string" }
					},
					{
						"name": "limit",
						"required": false,
						"in": "query",
						"description": "Max data points",
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "Timeseries data" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Get timeseries data",
				"tags": ["Analytics"]
			}
		},
		"/api/v1/analytics/{endpointId}/breakdown": {
			"get": {
				"description": "Returns analytics broken down by status code and HTTP method.",
				"operationId": "AnalyticsController_getBreakdown",
				"parameters": [
					{
						"name": "endpointId",
						"required": true,
						"in": "path",
						"description": "Endpoint UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "Breakdown data" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Get breakdown by status/method",
				"tags": ["Analytics"]
			}
		},
...
	},
	"info": {
		"title": "Proxy Server API",
		"description": "API documentation for the Proxy Server backend",
		"version": "1.0",
		"contact": { "name": "Proxy Server", "url": "", "email": "" },
		"license": { "name": "ISC", "url": "" }
	},
...
	"servers": [
		{ "url": "http://localhost:3000", "description": "development" },
		{ "url": "http://lvh.me:3000", "description": "production" }
	],
	"components": {
		"securitySchemes": {
			"Bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Enter JWT access token from /auth/sign-in, /auth/verify-email, or /auth/refresh",
				"in": "header",
				"name": "Authorization",
				"type": "http"
			}
		},
		"schemas": {
			"AuthUserSchema": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"description": "Unique user identifier (UUID)",
						"example": "550e8400-e29b-41d4-a716-446655440000",
						"format": "uuid"
					},
					"email": {
						"type": "string",
						"description": "User email address",
						"example": "user@example.com",
						"format": "email"
					},
					"name": {
						"type": "object",
						"description": "User display name",
						"example": "John Doe",
						"nullable": true
					}
				},
				"required": ["id", "email", "name"]
			},
			"AuthResponseSchema": {
				"type": "object",
				"properties": {
					"accessToken": {
						"type": "string",
						"description": "JWT access token for authenticated requests",
						"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
					},
					"user": {
						"description": "Authenticated user details",
						"allOf": [{ "$ref": "#/components/schemas/AuthUserSchema" }]
					}
				},
				"required": ["accessToken", "user"]
			},
			"SignUpDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"description": "User email address",
						"example": "user@example.com",
						"format": "email"
					},
					"password": {
						"type": "string",
						"description": "Password (minimum 8 characters)",
						"example": "SecureP@ssw0rd",
						"minLength": 8
					},
					"name": {
						"type": "object",
						"description": "Optional display name",
						"example": "John Doe"
					}
				},
				"required": ["email", "password"]
			},
			"VerifyEmailDto": {
				"type": "object",
				"properties": {
					"email": { "type": "string", "example": "user@example.com" },
					"code": {
						"type": "string",
						"example": "123456",
						"description": "6-digit code"
					}
				},
				"required": ["email", "code"]
			},
			"ResendVerificationDto": {
				"type": "object",
				"properties": {
					"email": { "type": "string", "example": "user@example.com" }
				},
				"required": ["email"]
			},
			"SignInDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"description": "User email address",
						"example": "user@example.com",
						"format": "email"
					},
					"password": {
						"type": "string",
						"description": "User password",
						"example": "SecureP@ssw0rd"
					}
				},
				"required": ["email", "password"]
			},
			"ForgotPasswordDto": {
				"type": "object",
				"properties": {
					"email": { "type": "string", "example": "user@example.com" }
				},
				"required": ["email"]
			},
			"ResetPasswordDto": {
				"type": "object",
				"properties": {
					"email": { "type": "string", "example": "user@example.com" },
					"code": { "type": "string", "example": "123456" },
					"newPassword": { "type": "string", "minLength": 8 }
				},
				"required": ["email", "code", "newPassword"]
			},
...
			"ErrorResponseSchema": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "number",
						"description": "HTTP status code",
						"example": 400,
						"minimum": 400,
						"maximum": 599
					},
					"error": {
						"type": "string",
						"description": "Error type or exception name",
						"example": "BadRequestException"
					},
					"message": {
						"description": "Human-readable error message(s)",
						"oneOf": [
							{ "type": "string", "example": "Validation failed" },
							{
								"type": "array",
								"items": { "type": "string" },
								"example": [
									"email must be an email",
									"password must be longer than 8 characters"
								]
							}
						]
					},
					"path": {
						"type": "string",
						"description": "Request path that triggered the error",
						"example": "/api/v1/auth/register"
					},
					"timestamp": {
						"type": "string",
						"description": "ISO 8601 timestamp when the error occurred",
						"example": "2025-03-18T12:00:00.000Z"
					}
				},
				"required": ["statusCode", "error", "message"]
			}
		}
	},
	"security": [{ "Bearer": [] }]
}